Address PR review comments for separated module registry
- Replace typing.Type with type alias to remove noqa comments
- Move test utilities to fme/core/registry/testing.py
- Add labels/encoding mismatch validation in SeparatedModule
- Use public .modules API in tests instead of internal properties
- Properly import timedelta in test file
